Remove version dependence, use Bukkit Messenger
