--- AppleSkinSpigotCore/src/main/java/com/jmatt/appleskinspigot/craftplayer/BukkitCraftPlayer.java
@@ -1,10 +1,0 @@
-package com.jmatt.appleskinspigot.craftplayer;
-
-import net.minecraft.network.protocol.game.PacketPlayOutCustomPayload;
-import org.bukkit.entity.Player;
-
-public interface BukkitCraftPlayer {
-
-    void sendPacket(Player player, PacketPlayOutCustomPayload packet);
-
-}
