--- AppleSkinSpigotPlugin/src/main/java/com/jmatt/appleskinspigot/UnsupportedMinecraftVersionException.java
@@ -1,3 +1,0 @@
-package com.jmatt.appleskinspigot;
-
-public class UnsupportedMinecraftVersionException extends Exception {}
