--- AppleSkinSpigotPlugin/src/main/resources/plugin.yml
@@ -1,4 +1,0 @@
-name: AppleCoreSpigot
-version: 1.1.1
-author: jmatt
-main: com.jmatt.appleskinspigot.AppleSkinSpigotPlugin
