--- AppleSkinSpigot_1_18/src/test/java/com/jmatt/appleskinspigot/craftplayer/BukkitCraftPlayer_1_18_Test.java
@@ -1,38 +1,0 @@
-package com.jmatt.appleskinspigot.craftplayer;
-
-import net.minecraft.network.protocol.game.PacketPlayOutCustomPayload;
-import net.minecraft.server.level.EntityPlayer;
-import net.minecraft.server.network.PlayerConnection;
-import org.bukkit.entity.Player;
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.mockito.Mock;
-import org.mockito.Mockito;
-import org.mockito.junit.MockitoJUnitRunner;
-
-import static org.mockito.Mockito.times;
-import static org.mockito.Mockito.verify;
-import static org.mockito.Mockito.when;
-
-@RunWith(MockitoJUnitRunner.class)
-public class BukkitCraftPlayer_1_18_Test {
-
-    @Mock
-    EntityPlayer entityPlayerMock;
-    @Mock
-    PlayerConnection playerConnectionMock;
-    @Mock
-    PacketPlayOutCustomPayload packetMock;
-
-    Player playerMock;
-
-    @Test
-    public void sendPacket1_18_1SendsPacket() {
-        playerMock = Mockito.mock(org.bukkit.craftbukkit.v1_18_R1.entity.CraftPlayer.class);
-        when(((org.bukkit.craftbukkit.v1_18_R1.entity.CraftPlayer)playerMock).getHandle()).thenReturn(entityPlayerMock);
-        entityPlayerMock.b = playerConnectionMock;
-        new BukkitCraftPlayer_1_18().sendPacket(playerMock, packetMock);
-        verify(playerConnectionMock, times(1)).a(packetMock);
-    }
-
-}
