+++ src/main/resources/plugin.yml
@@ -1,0 +1,4 @@
+name: AppleCoreSpigot
+version: 1.1.1
+author: jmatt
+main: com.jmatt.appleskinspigot.AppleSkinSpigotPlugin
