+++ src/test/java/com/jmatt/appleskinspigot/LoginListenerTest.java
